use resources path to store the customer invoice in the target folder (create after compile)

+++ src/main/java/com/ecommerce/ecommApp/billGenerator/dto/InvoiceDetails.java
@@ -32,9 +32,9 @@ public class InvoiceDetails {
     private String productImage;
 
     @JsonProperty(value = "price", required = true)
-    private Long price;
+    private Float price;
 
     @JsonProperty(value = "total_amount", required = true)
-    private Long totalAmount;
+    private Double totalAmount;
 
 }
